DRY out TfsConnectionFactory
