+++ src/Qwiq.Core.Rest/TfsConnectionFactory.cs
@@ -1,13 +1,11 @@
 using System;
 
-using Microsoft.Qwiq.Credentials;
-using Microsoft.TeamFoundation.Build.WebApi;
 using Microsoft.VisualStudio.Services.Common;
 using Microsoft.VisualStudio.Services.WebApi;
 
 namespace Microsoft.Qwiq.Client.Rest
 {
-    public class TfsConnectionFactory : Qwiq.TfsConnectionFactory
+    public class TfsConnectionFactory : TfsConnectionFactory<ITeamProjectCollection>
     {
         public static readonly ITfsConnectionFactory Default = Nested.Instance;
 
@@ -15,39 +13,12 @@ namespace Microsoft.Qwiq.Client.Rest
         {
         }
 
-        /// <inheritdoc />
-        public override ITeamProjectCollection Create(AuthenticationOptions options)
-        {
-            if (options == null) throw new ArgumentNullException(nameof(options));
-            var credentials = options.Credentials;
-
-            foreach (var credential in credentials)
-            {
-                try
-                {
-                    var tfsNative = ConnectToTfsCollection(options.Uri, credential);
-                    var tfsProxy = tfsNative.AsProxy();
-                    options.Notifications.AuthenticationSuccess(new AuthenticationSuccessNotification(credential, tfsProxy));
-
-                    return tfsProxy;
-                }
-                catch (Exception e)
-                {
-                    options.Notifications.AuthenticationFailed(new AuthenticationFailedNotification(credential, e));
-                }
-            }
-
-            var nocreds = new AccessDeniedException("Invalid credentials");
-            options.Notifications.AuthenticationFailed(new AuthenticationFailedNotification(null, nocreds));
-            throw nocreds;
-        }
-
-        private static VssConnection ConnectToTfsCollection(Uri endpoint, VssCredentials credentials)
+        protected override ITeamProjectCollection ConnectToTfsCollection(Uri endpoint, VssCredentials credentials)
         {
             var tfsServer = new VssConnection(endpoint, credentials);
             tfsServer.ConnectAsync(VssConnectMode.Automatic).GetAwaiter().GetResult();
             if (!tfsServer.HasAuthenticated) throw new InvalidOperationException("Could not connect.");
-            return tfsServer;
+            return tfsServer.AsProxy();
         }
 
         // ReSharper disable ClassNeverInstantiated.Local
@@ -56,6 +27,7 @@ namespace Microsoft.Qwiq.Client.Rest
         {
             // ReSharper disable MemberHidesStaticFromOuterClass
             internal static readonly TfsConnectionFactory Instance = new TfsConnectionFactory();
+
             // ReSharper restore MemberHidesStaticFromOuterClass
 
             // Explicit static constructor to tell C# compiler
