+++ src/Qwiq.Core.Soap/TfsConnectionFactory.cs
@@ -1,12 +1,10 @@
 using System;
 
-using Microsoft.Qwiq.Credentials;
-using Microsoft.TeamFoundation.Build.Client;
 using Microsoft.VisualStudio.Services.Common;
 
 namespace Microsoft.Qwiq.Client.Soap
 {
-    public class TfsConnectionFactory : Qwiq.TfsConnectionFactory
+    public class TfsConnectionFactory : TfsConnectionFactory<ITeamProjectCollection>
     {
         public static readonly ITfsConnectionFactory Default = Nested.Instance;
 
@@ -15,38 +13,11 @@ namespace Microsoft.Qwiq.Client.Soap
         }
 
         /// <inheritdoc />
-        public override ITeamProjectCollection Create(AuthenticationOptions options)
-        {
-            if (options == null) throw new ArgumentNullException(nameof(options));
-            var credentials = options.Credentials;
-
-            foreach (var credential in credentials)
-            {
-                try
-                {
-                    var tfsNative = ConnectToTfsCollection(options.Uri, credential);
-                    var tfsProxy = tfsNative.AsProxy();
-
-                    options.Notifications.AuthenticationSuccess(new AuthenticationSuccessNotification(credential, tfsProxy));
-
-                    return tfsProxy;
-                }
-                catch (Exception e)
-                {
-                    options.Notifications.AuthenticationFailed(new AuthenticationFailedNotification(credential, e));
-                }
-            }
-
-            var nocreds = new AccessDeniedException("Invalid credentials");
-            options.Notifications.AuthenticationFailed(new AuthenticationFailedNotification(null, nocreds));
-            throw nocreds;
-        }
-
-        private static TeamFoundation.Client.TfsTeamProjectCollection ConnectToTfsCollection(Uri endpoint, VssCredentials credentials)
+        protected override ITeamProjectCollection ConnectToTfsCollection(Uri endpoint, VssCredentials credentials)
         {
             var tfsServer = new TeamFoundation.Client.TfsTeamProjectCollection(endpoint, credentials);
             tfsServer.EnsureAuthenticated();
-            return tfsServer;
+            return tfsServer.AsProxy();
         }
 
         // ReSharper disable ClassNeverInstantiated.Local
@@ -55,6 +26,7 @@ namespace Microsoft.Qwiq.Client.Soap
         {
             // ReSharper disable MemberHidesStaticFromOuterClass
             internal static readonly TfsConnectionFactory Instance = new TfsConnectionFactory();
+
             // ReSharper restore MemberHidesStaticFromOuterClass
 
             // Explicit static constructor to tell C# compiler
