+++ src/Qwiq.Core/ITfsConnectionFactory.cs
@@ -1,0 +1,9 @@
+﻿using Microsoft.Qwiq.Credentials;
+
+namespace Microsoft.Qwiq
+{
+    public interface ITfsConnectionFactory
+    {
+        ITeamProjectCollection Create(AuthenticationOptions options);
+    }
+}
