+++ src/Qwiq.Core/Qwiq.Core.csproj
@@ -64,6 +64,7 @@
     <Compile Include="..\AssemblyInfo.Common.cs">
       <Link>Properties\AssemblyInfo.Common.cs</Link>
     </Compile>
+    <Compile Include="AccessDeniedException.cs" />
     <Compile Include="BaseLinkType.cs" />
     <Compile Include="ClientType.cs" />
     <Compile Include="Comparer.cs" />
@@ -143,6 +144,7 @@
     <Compile Include="IRevisionInternal.cs" />
     <Compile Include="ITeamFoundationIdentity.cs" />
     <Compile Include="ITeamFoundationIdentity.Extensions.cs" />
+    <Compile Include="ITfsConnectionFactory.cs" />
     <Compile Include="ITfsTeamProjectCollection.cs" />
     <Compile Include="ITypeParser.cs" />
     <Compile Include="IWorkItem.cs" />
